# public/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"IRoute": {
				"properties": {
					"route": {
						"items": {
							"items": {
								"type": "number",
								"format": "double"
							},
							"type": "array"
						},
						"type": "array"
					},
					"distance": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"route",
					"distance"
				],
				"type": "object",
				"additionalProperties": false
			},
			"DeliveryRequest": {
				"properties": {
					"delivery_ids": {
						"items": {
							"type": "number",
							"format": "double"
						},
						"type": "array"
					}
				},
				"required": [
					"delivery_ids"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "prueba-minery-express-typescript",
		"version": "1.0.0",
		"contact": {}
	},
	"paths": {
...
		"/routes": {
			"post": {
				"operationId": "CreateRoute",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRoute"
								}
							}
						}
					},
					"400": {
						"description": "Error creating route"
					}
				},
				"tags": [
					"Routes"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/DeliveryRequest"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}